Change `ValidatorSetProof.eventBlock` type uint256 -> uint64

+++ contracts/contracts/checks/CheckAura.sol
@@ -41,7 +41,7 @@ contract CheckAura is Initializable {
     struct ValidatorSetProof {
         bytes[] receiptProof;
         ValidatorSetChange[] changes;
-        uint eventBlock;
+        uint64 eventBlock;
     }
 
     struct AuraProof {
